requirements.txt
alabaster==0.7.12
asgiref==3.2.3
asn1crypto==0.24.0
awscli==1.16.263
Babel==2.8.0
botocore==1.12.253
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
cloudinary==1.17.0
colorama==0.4.1
cryptography==2.7
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.12
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-cloudinary-storage==0.2.3
django-cors-headers==2.2.0
django-filter==1.1.0
django-imagekit==4.0.2
django-import-export==1.0.1
django-python3-ldap==0.11.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.9.1
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.15.2
drf-generators==0.3.0
et-xmlfile==1.0.1
feedparser==5.2.1
gitsome==0.8.4
gnupg==2.3.1
gunicorn==19.9.0
idna==2.8
imagesize==1.2.0
jdcal==1.4
Jinja2==2.11.2
jmespath==0.9.4
ldap3==2.3
MarkupSafe==1.1.1
mock==4.0.2
node==0.9.23
nodejs==0.1.1
numpydoc==0.9.2
odfpy==1.3.6
odict==1.6.2
openpyxl==2.5.4
optional-django==0.1.0
packaging==20.3
pilkit==2.0
Pillow==6.2.0
pipenv==2018.11.26
plumber==1.5
ply==3.11
post==2019.4.13
prompt-toolkit==2.0.10
psutil==5.6.3
psycopg2-binary==2.8.5
pyasn1==0.3.2
pyasn1-modules==0.2.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.6.4
pyparsing==2.4.7
python-dateutil==2.8.0
python-decouple==3.1
python-ldap==3.1.0
pytz==2016.10
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.11.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.25.6
virtualenv==16.6.1
virtualenv-clone==0.5.3
wcwidth==0.1.7
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.1.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
